backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.0.350
langchain-google-genai==1.0.10
google-generativeai>=0.7.0,<0.8.0
openai==1.3.8
anthropic==0.7.8
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-dotenv==1.0.0
python-multipart==0.0.6
reportlab==4.0.7
jinja2==3.1.2
httpx==0.25.2
redis==5.0.1
celery==5.3.4
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9